+++ src/main/java/com/github/alexthe666/iceandfire/client/render/tile/RenderDreadSpawner.java
@@ -8,7 +8,6 @@ import net.minecraft.client.renderer.Vector3f;
 import net.minecraft.client.renderer.tileentity.TileEntityRenderer;
 import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
 import net.minecraft.entity.Entity;
-import net.minecraft.tileentity.MobSpawnerTileEntity;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.spawner.AbstractSpawner;
 
@@ -26,13 +25,13 @@ public class RenderDreadSpawner<T extends TileEntityDreadSpawner> extends TileEn
         if (entity != null) {
             float f = 0.53125F;
             float f1 = Math.max(entity.getWidth(), entity.getHeight());
-            if ((double)f1 > 1.0D) {
+            if ((double) f1 > 1.0D) {
                 f /= f1;
             }
 
-            matrixStackIn.translate(0.0D, (double)0.4F, 0.0D);
-            matrixStackIn.rotate(Vector3f.YP.rotationDegrees((float) MathHelper.lerp((double)partialTicks, abstractspawner.getPrevMobRotation(), abstractspawner.getMobRotation()) * 10.0F));
-            matrixStackIn.translate(0.0D, (double)-0.2F, 0.0D);
+            matrixStackIn.translate(0.0D, 0.4F, 0.0D);
+            matrixStackIn.rotate(Vector3f.YP.rotationDegrees((float) MathHelper.lerp(partialTicks, abstractspawner.getPrevMobRotation(), abstractspawner.getMobRotation()) * 10.0F));
+            matrixStackIn.translate(0.0D, -0.2F, 0.0D);
             matrixStackIn.rotate(Vector3f.XP.rotationDegrees(-30.0F));
             matrixStackIn.scale(f, f, f);
             Minecraft.getInstance().getRenderManager().renderEntityStatic(entity, 0.0D, 0.0D, 0.0D, 0.0F, partialTicks, matrixStackIn, bufferIn, combinedLightIn);
