+++ src/main/java/com/github/alexthe666/iceandfire/compat/jei/firedragonforge/FireDragonForgeDrawable.java
@@ -34,10 +34,10 @@ public class FireDragonForgeDrawable implements IDrawable {
         Tessellator tessellator = Tessellator.getInstance();
         BufferBuilder bufferbuilder = tessellator.getBuffer();
         bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-        bufferbuilder.pos((double) (x + 0), (double) (y + height), (double) 0).tex((float) ((float) (textureX + 0) * 0.00390625F), (float) ((float) (textureY + height) * 0.00390625F)).endVertex();
-        bufferbuilder.pos((double) (x + width), (double) (y + height), (double) 0).tex((float) ((float) (textureX + width) * 0.00390625F), (float) ((float) (textureY + height) * 0.00390625F)).endVertex();
-        bufferbuilder.pos((double) (x + width), (double) (y + 0), (double) 0).tex((float) ((float) (textureX + width) * 0.00390625F), (float) ((float) (textureY + 0) * 0.00390625F)).endVertex();
-        bufferbuilder.pos((double) (x + 0), (double) (y + 0), (double) 0).tex((float) ((float) (textureX + 0) * 0.00390625F), (float) ((float) (textureY + 0) * 0.00390625F)).endVertex();
+        bufferbuilder.pos(x + 0, y + height, 0).tex((float) (textureX + 0) * 0.00390625F, (float) (textureY + height) * 0.00390625F).endVertex();
+        bufferbuilder.pos(x + width, y + height, 0).tex((float) (textureX + width) * 0.00390625F, (float) (textureY + height) * 0.00390625F).endVertex();
+        bufferbuilder.pos(x + width, y + 0, 0).tex((float) (textureX + width) * 0.00390625F, (float) (textureY + 0) * 0.00390625F).endVertex();
+        bufferbuilder.pos(x + 0, y + 0, 0).tex((float) (textureX + 0) * 0.00390625F, (float) (textureY + 0) * 0.00390625F).endVertex();
         tessellator.draw();
     }
 }
